crowdfunding_django/requirements.txt
Django==5.0.3
psycopg2-binary==2.9.9
Pillow==10.2.0
python-decouple==3.8
whitenoise==6.6.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2